obj_detection_django/requirements.txt
absl-py==1.0.0
anyio==3.4.0
asgiref==3.4.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.0.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
docopt==0.6.2
drf-yasg==1.20.0
fastapi==0.70.1
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.0
oauthlib==3.1.1
openapi-codec==1.3.2
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pipreqs==0.4.11
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-multipart==0.0.5
pytube==11.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scipy==1.7.3
seaborn==0.11.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.16.0
Werkzeug==2.0.2
wrapt==1.13.3
yarg==0.1.9
zipp==3.7.0
